JWT Implementation, validation added
