+++ src/main/java/com/Energy/BasicSpringAPI/controller/AuthController.java
@@ -2,6 +2,8 @@ package com.Energy.BasicSpringAPI.controller;
 
 import com.Energy.BasicSpringAPI.DTO.UserDto;
 import com.Energy.BasicSpringAPI.entity.UserEntity;
+import com.Energy.BasicSpringAPI.service.AuthService;
+import com.Energy.BasicSpringAPI.service.AuthenticationFilter;
 import com.Energy.BasicSpringAPI.service.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -22,6 +24,12 @@ import java.util.StringTokenizer;
 @RequestMapping("auth")
 public class AuthController {
     @Autowired
+    private AuthService authService;
+
+    @Autowired
+    private AuthenticationFilter authenticationFilter;
+
+    @Autowired
     private UserService userService;
 
     @PermitAll
@@ -34,19 +42,19 @@ public class AuthController {
         System.out.println(userName);
         System.out.println(password);
 
-        Optional<UserEntity> user = userService.getUser(userName, password);
+        Optional<UserEntity> user = authService.getUser(userName, password);
         if (user == null){
-            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+            return new ResponseEntity<>("The email or password is wrong", HttpStatus.UNAUTHORIZED);
         }
         else {
             String userId = Long.toString(user.get().id);
-            String token = userService.createJWT(userId, user.get().email, user.get().username, -1);
-            if (userService.validateToken(token)) {
+            String token = authenticationFilter.createJWT(userId, user.get().email, user.get().username, -1);
+            if (authenticationFilter.validateToken(token)) {
                 return new ResponseEntity<>(token, HttpStatus.OK);
 
             }
             else {
-                return new ResponseEntity<>("The email or password ir wrong", HttpStatus.TOO_EARLY);
+                return new ResponseEntity<>("The email or password is wrong", HttpStatus.UNAUTHORIZED);
             }
         }
     }
@@ -54,17 +62,14 @@ public class AuthController {
 
     @RequestMapping(value = "/verifyToken", method = RequestMethod.POST)
     public ResponseEntity VerifyTOken(HttpServletResponse response, @RequestBody String body) throws IOException, SQLException, URISyntaxException, NoSuchAlgorithmException {
-        System.out.println(body);
         final StringTokenizer tokenizer = new StringTokenizer(body, ":");
         final String token = tokenizer.nextToken();
-        System.out.println(token);
 
-        if (userService.validateToken(token)) {
-            return new ResponseEntity<>(token, HttpStatus.OK);
-
+        if (authenticationFilter.validateToken(token)) {
+            return new ResponseEntity<>("The token is valid", HttpStatus.OK);
         }
         else {
-            return new ResponseEntity<>("The token is invalid", HttpStatus.TOO_EARLY);
+            return new ResponseEntity<>("The token is invalid", HttpStatus.BAD_REQUEST);
         }
     }
