+++ src/main/java/com/Energy/BasicSpringAPI/service/AuthenticationFilter.java
@@ -1,0 +1,105 @@
+package com.Energy.BasicSpringAPI.service;
+
+import io.jsonwebtoken.*;
+import org.bouncycastle.jcajce.BCFKSLoadStoreParameter;
+import org.springframework.stereotype.Component;
+import org.springframework.stereotype.Service;
+
+import javax.crypto.spec.SecretKeySpec;
+import javax.xml.bind.DatatypeConverter;
+import java.nio.charset.StandardCharsets;
+import java.security.Key;
+import java.security.MessageDigest;
+import java.security.NoSuchAlgorithmException;
+import java.sql.Date;
+import java.time.ZonedDateTime;
+
+@Component
+public class AuthenticationFilter {
+
+    public static String doHashing (String password) throws NoSuchAlgorithmException {
+        final MessageDigest digest = MessageDigest.getInstance("SHA3-256");
+        final byte[] hashbytes = digest.digest(
+                password.getBytes(StandardCharsets.UTF_8));
+        String sha3Hex = bytesToHex(hashbytes);
+
+        return sha3Hex;
+    }
+    private static String bytesToHex(byte[] hash) {
+        StringBuilder hexString = new StringBuilder(2 * hash.length);
+        for (int i = 0; i < hash.length; i++) {
+            String hex = Integer.toHexString(0xff & hash[i]);
+            if(hex.length() == 1) {
+                hexString.append('0');
+            }
+            hexString.append(hex);
+        }
+        return hexString.toString();
+    }
+
+    //    public UserEntity getUserFromToken(String token) {
+//        Claims decoded = decodeJWT(token);
+//
+//        String id = decoded.getId();
+//
+//        UserEntity u = getUser(parseInt(id));
+//
+//        return u;
+//    }
+    public static String SECRET_KEY = "oeRaYY";
+
+    public String createJWT(String id, String issuer, String subject, long ttlMillis) {
+
+        //The JWT signature algorithm we will be using to sign the token
+        SignatureAlgorithm signatureAlgorithm = SignatureAlgorithm.HS256;
+
+        long nowMillis = System.currentTimeMillis();
+        Date now = new Date(nowMillis);
+
+        //We will sign our JWT with our ApiKey secret
+        byte[] apiKeySecretBytes = DatatypeConverter.parseBase64Binary(SECRET_KEY);
+        Key signingKey = new SecretKeySpec(apiKeySecretBytes, signatureAlgorithm.getJcaName());
+
+        //Let's set the JWT Claims
+        JwtBuilder builder = Jwts.builder().setId(id)
+                .setIssuedAt(now)
+                .setSubject(subject)
+                .setIssuer(issuer)
+                .setExpiration(Date.from(ZonedDateTime.now().plusMinutes(1).toInstant()))
+                .signWith(signatureAlgorithm, signingKey);
+
+        //Builds the JWT and serializes it to a compact, URL-safe string
+
+        return builder.compact();
+    }
+
+    public boolean validateToken(String token) {
+        boolean validation = false;
+        try {
+            Jws<Claims> claims  = Jwts.parser().setSigningKey(SECRET_KEY).parseClaimsJws(token);
+            if(!claims.getBody().getExpiration().before(new java.util.Date())){
+                validation = true;
+            }
+            else
+            {
+                validation = false;
+            }
+
+        } catch (ExpiredJwtException e) {
+            System.out.println(" Token expired ");
+        } catch(Exception e){
+            System.out.println(" Some other exception in JWT parsing ");
+        }
+        return validation;
+
+    }
+    public Claims decodeJWT(String jwt) {
+
+        //This line will throw an exception if it is not a signed JWS (as expected)
+        Claims claims = Jwts.parser()
+                .setSigningKey(DatatypeConverter.parseBase64Binary(SECRET_KEY))
+                .parseClaimsJws(jwt).getBody();
+        return claims;
+    }
+
+}
